+++ demo/src/app/shared/html-sanitizer.pipe.ts
@@ -1,4 +1,4 @@
-import { Pipe, PipeTransform } from '@angular/core';
+import {Pipe, PipeTransform} from '@angular/core';
 import {DomSanitizer} from '@angular/platform-browser';
 
 @Pipe({
@@ -6,7 +6,8 @@ import {DomSanitizer} from '@angular/platform-browser';
 })
 export class HtmlSanitizerPipe implements PipeTransform {
 
-  constructor(private _sanitizer: DomSanitizer){}
+  constructor(private _sanitizer: DomSanitizer) {
+  }
 
   transform(value: any): any {
     return this._sanitizer.bypassSecurityTrustHtml(value);