+++ CHANGELOG.md
@@ -1,1 +1,17 @@
 # Changelog
+
+## 0.6.0 beta
+
+ - Add popup animation for gallery modal
+ - Remove incorrect slide animation
+ - Make gestures optional, closes [#2](https://github.com/MurhafSousli/ng-gallery/issues/2)
+ - Remove thumbnail vertical position (right and left) positions, closes [#3](https://github.com/MurhafSousli/ng-gallery/issues/3)
+
+## 0.5.2 beta
+
+ - (feat) gestures support
+ - (refactor) gallery config 
+
+## 0.5.0 beta
+
+ - Initial release